+++ model_builder/domain/services/__init__.py
@@ -7,6 +7,10 @@ from model_builder.domain.services.system_validation_service import SystemValida
 from model_builder.domain.services.emissions_calculation_service import EmissionsCalculationService, EmissionsResult
 from model_builder.domain.services.edit_service import EditService, EditResult
 from model_builder.domain.services.progressive_import_service import ProgressiveImportService
+from model_builder.domain.services.empty_model import is_empty_model
+from model_builder.domain.services.template_catalog_service import (
+    CatalogEntry, CatalogGroup, SCRATCH_ID, build_template_catalog, get_template_system_data,
+)
 
 __all__ = [
     "ObjectLinkingService",
@@ -14,4 +18,6 @@ __all__ = [
     "EmissionsCalculationService", "EmissionsResult",
     "EditService", "EditResult",
     "ProgressiveImportService",
+    "is_empty_model",
+    "CatalogEntry", "CatalogGroup", "SCRATCH_ID", "build_template_catalog", "get_template_system_data",
 ]